Move helpers to utilities to avoid name collision
